+++ src/Maestro/API/V1/Addresses.hs
@@ -7,13 +7,13 @@ import           Servant.API.Generic
 
 data AddressesAPI route = AddressesAPI
   {
-    _decodeAddress
+    decodeAddress
       :: route
       :- Capture "address" (TaggedText AddressToDecode)
       :> "decode"
       :> Get '[JSON] AddressInfo
 
-  , _addressesUtxos
+  , addressesUtxos
       :: route
       :- "utxos"
       :> QueryParam "resolve_datums" Bool
@@ -22,11 +22,22 @@ data AddressesAPI route = AddressesAPI
       :> ReqBody '[JSON] [Bech32StringOf Address]
       :> Post '[JSON] PaginatedUtxoWithSlot
 
-  , _addressUtxoRefs
+  , addressUtxoRefs
       :: route
       :- Capture "address" (Bech32StringOf Address)
       :> "utxo_refs"
       :> Pagination
       :> Get '[JSON] PaginatedOutputReferenceObject
+
+  , paymentCredentialUtxos
+      :: route
+      :- "cred"
+      :> Capture "credential" (Bech32StringOf PaymentCredentialAddress)
+      :> "utxos"
+      :> QueryParam "resolve_datums" Bool
+      :> QueryParam "with_cbor" Bool
+      -- TODO: Support for more query parameters.
+      :> Pagination
+      :> Get '[JSON] PaginatedUtxoWithSlot
 
   } deriving (Generic)