+++ src/Maestro/Client/Env.hs
@@ -1,15 +1,15 @@
 module Maestro.Client.Env
-  (
-    MaestroEnv (..)
+  ( MaestroEnv (..)
   , MaestroNetwork (..)
   , MaestroApiVersion (..)
   , mkMaestroEnv
+  , defaultBackoff
   ) where
 
+import           Data.Functor            ((<&>))
 import           Data.Text               (Text)
 import           Network.HTTP.Client     (newManager)
 import           Network.HTTP.Client.TLS (tlsManagerSettings)
-
 import qualified Servant.Client          as Servant
 
 type MaestroToken = Text
@@ -35,8 +35,14 @@ instance SingMaestroApiVersionI 'V0 where singMaestroApiVersion = SingV0
 instance SingMaestroApiVersionI 'V1 where singMaestroApiVersion = SingV1
 
 data MaestroEnv (v :: MaestroApiVersion) = MaestroEnv
-  { _maeClientEnv :: !Servant.ClientEnv
-  , _maeToken     :: !MaestroToken
+  { maeClientEnv :: !Servant.ClientEnv
+  , maeToken     :: !MaestroToken
+  , maeBaseDelay :: !(Maybe Int)
+  -- ^ Base delay in microseconds to use with jitter backoff.
+  -- https://hackage.haskell.org/package/retry-0.9.3.1/docs/Control-Retry.html#v:exponentialBackoff
+  , maeMaxDelay  :: !(Maybe Int)
+  -- ^ Maximum waiting time in microseconds.
+  -- https://hackage.haskell.org/package/retry-0.9.3.1/docs/Control-Retry.html#v:limitRetriesByCumulativeDelay
   }
 
 data MaestroNetwork = Mainnet | Preprod | Preview
@@ -46,13 +52,29 @@ maestroBaseUrl Preview v = "https://preview.gomaestro-api.org/" <> show v
 maestroBaseUrl Preprod v = "https://preprod.gomaestro-api.org/" <> show v
 maestroBaseUrl Mainnet v = "https://mainnet.gomaestro-api.org/" <> show v
 
-mkMaestroEnv :: forall (apiVersion :: MaestroApiVersion). SingMaestroApiVersionI apiVersion => MaestroToken -> MaestroNetwork -> IO (MaestroEnv apiVersion)
-mkMaestroEnv token nid = do
+mkMaestroEnv
+  :: forall (apiVersion :: MaestroApiVersion).
+  ( SingMaestroApiVersionI apiVersion
+  ) =>
+  MaestroToken ->
+  MaestroNetwork ->
+  Maybe (Int, Int) ->
+  IO (MaestroEnv apiVersion)
+mkMaestroEnv token nid mbDelays = do
   clientEnv <- servantClientEnv $ maestroBaseUrl nid (fromSingMaestroApiVersion $ singMaestroApiVersion @apiVersion)
-  pure $ MaestroEnv { _maeClientEnv = clientEnv, _maeToken = token }
+  pure $ MaestroEnv
+    { maeClientEnv = clientEnv
+    , maeToken = token
+    , maeBaseDelay = mbDelays <&> fst
+    , maeMaxDelay = mbDelays <&> snd
+    }
 
 servantClientEnv :: String -> IO Servant.ClientEnv
 servantClientEnv url = do
   baseUrl <- Servant.parseBaseUrl url
   manager <- newManager tlsManagerSettings
   pure $ Servant.mkClientEnv manager baseUrl
+
+-- | Base delay & maximum waiting in microseconds corresponding to `maeBaseDelay`, `maeMaxDelay` respectively. See description of `maeBaseDelay` & `maeMaxDelay` for more information.
+defaultBackoff :: Maybe (Int, Int)
+defaultBackoff = Just (1_000_000, 20_000_000)