+++ src/Maestro/Client/V1/Datum.hs
@@ -4,7 +4,7 @@ module Maestro.Client.V1.Datum
   ( getDatumByHash
   ) where
 
-import           Maestro.API.V1         (_datum)
+import           Maestro.API.V1         (datum)
 import           Maestro.API.V1.Datum
 import           Maestro.Client.Env
 import           Maestro.Client.V1.Core
@@ -13,8 +13,8 @@ import           Servant.API.Generic
 import           Servant.Client
 
 datumClient :: MaestroEnv 'V1 -> DatumAPI (AsClientT IO)
-datumClient = fromServant . _datum . apiV1Client
+datumClient = fromServant . datum . apiV1Client
 
 -- | Get information about the datum from it's hash.
 getDatumByHash :: MaestroEnv 'V1 -> HexStringOf DatumHash -> IO TimestampedDatum
-getDatumByHash = _datumByHash . datumClient
+getDatumByHash = datumByHash . datumClient