+++ src/Maestro/Client/V1/TxManager.hs
@@ -6,7 +6,7 @@ where
 
 import qualified Data.ByteString          as BS
 import           Data.Text                (Text)
-import           Maestro.API.V1           (_txManager)
+import           Maestro.API.V1           (txManager)
 import           Maestro.API.V1.TxManager
 import           Maestro.Client.Env
 import           Maestro.Client.V1.Core
@@ -14,7 +14,7 @@ import           Servant.API.Generic
 import           Servant.Client
 
 txClient :: MaestroEnv 'V1 -> TxManagerAPI (AsClientT IO)
-txClient = fromServant . _txManager . apiV1Client
+txClient = fromServant . txManager . apiV1Client
 
 -- | Submit a signed and serialized transaction to the network. A transaction submited with this endpoint will be monitored by Maestro.
 submitAndMonitorTx ::
@@ -23,7 +23,7 @@ submitAndMonitorTx ::
   -- | CBOR encoded Transaction.
   BS.ByteString ->
   IO Text
-submitAndMonitorTx = _monitoredTxSubmit . txClient
+submitAndMonitorTx = monitoredTxSubmit . txClient
 
 -- | Submit a signed and serialized transaction to the network. A transaction submited with this endpoint will be [turbo submitted & monitored by Maestro](https://docs.gomaestro.org/docs/API%20reference/Transaction%20Manager/tx-manager-turbo-submit).
 turboSubmitAndMonitorTx ::
@@ -32,4 +32,4 @@ turboSubmitAndMonitorTx ::
   -- | CBOR encoded Transaction.
   BS.ByteString ->
   IO Text
-turboSubmitAndMonitorTx = _monitoredTurboTxSubmit . txClient
+turboSubmitAndMonitorTx = monitoredTurboTxSubmit . txClient