+++ src/Maestro/Types/V1.hs
@@ -1,7 +1,9 @@
 -- | Module exporting all available types of this specific Maestro-API version.
 
 module Maestro.Types.V1
-  ( module Maestro.Types.V1.Addresses
+  ( module Maestro.Types.V1.Accounts
+  , module Maestro.Types.V1.Addresses
+  , module Maestro.Types.V1.Blocks
   , module Maestro.Types.V1.Common
   , module Maestro.Types.V1.Datum
   , module Maestro.Types.V1.General
@@ -9,7 +11,9 @@ module Maestro.Types.V1
   , module Maestro.Types.V1.Transactions
   ) where
 
+import           Maestro.Types.V1.Accounts
 import           Maestro.Types.V1.Addresses
+import           Maestro.Types.V1.Blocks
 import           Maestro.Types.V1.Common
 import           Maestro.Types.V1.Datum
 import           Maestro.Types.V1.General