+++ src/Maestro/Types/V1/Addresses.hs
@@ -34,16 +34,16 @@ data PaymentCredKind = PCKKey | PCKScript
 
 -- | Payment credential, the payment part of a Cardano address.
 data PaymentCredential = PaymentCredential
-  { _paymentCredentialBech32:: !(Bech32StringOf PaymentCredentialAddress)
+  { paymentCredentialBech32:: !(Bech32StringOf PaymentCredentialAddress)
   -- ^ Bech32-encoding of the credential key hash or script hash.
-  , _paymentCredentialHex    :: !(HexStringOf PaymentCredentialAddress)
+  , paymentCredentialHex    :: !(HexStringOf PaymentCredentialAddress)
   -- ^ Hex-encoding of the script or key credential.
-  , _paymentCredentialKind   :: !PaymentCredKind
+  , paymentCredentialKind   :: !PaymentCredKind
   -- ^ See `PaymentCredKind`.
   }
   deriving stock (Show, Eq, Ord, Generic)
   deriving (FromJSON, ToJSON)
-  via CustomJSON '[FieldLabelModifier '[StripPrefix "_paymentCredential", CamelToSnake]] PaymentCredential
+  via CustomJSON '[FieldLabelModifier '[StripPrefix "paymentCredential", CamelToSnake]] PaymentCredential
 
 -- | Denotes kind of a staking credential.
 data StakingCredKind = SCKKey | SCKScript | SCKPointer
@@ -57,77 +57,77 @@ newtype CertIndex = CertIndex Natural
 
 -- | In an address, a chain pointer refers to a point of the chain containing a stake key registration certificate. A point is identified by 3 coordinates, as listed in the type.
 data ChainPointer = ChainPointer
-  { _chainPointerSlot      :: !SlotNo
+  { chainPointerSlot      :: !SlotNo
   -- ^ An absolute slot number.
-  , _chainPointerTxIndex   :: !TxIndex
+  , chainPointerTxIndex   :: !TxIndex
   -- ^ A transaction index (within that slot).
-  , _chainPointerCertIndex :: !CertIndex
+  , chainPointerCertIndex :: !CertIndex
   -- ^ A (delegation) certificate index (within that transaction).
   }
   deriving stock (Show, Eq, Ord, Generic)
   deriving (FromJSON, ToJSON)
-  via CustomJSON '[FieldLabelModifier '[StripPrefix "_chainPointer", CamelToSnake]] ChainPointer
+  via CustomJSON '[FieldLabelModifier '[StripPrefix "chainPointer", CamelToSnake]] ChainPointer
 
 -- | Payment credential, the payment part of a Cardano address.
 data StakingCredential = StakingCredential
-  { _stakingCredentialBech32:: !(Maybe (Bech32StringOf StakingCredentialAddress))
+  { stakingCredentialBech32:: !(Maybe (Bech32StringOf StakingCredentialAddress))
   -- ^ Bech32-encoding of the credential key hash or script hash.
-  , _stakingCredentialHex   :: !(Maybe (HexStringOf StakingCredentialAddress))
+  , stakingCredentialHex   :: !(Maybe (HexStringOf StakingCredentialAddress))
   -- ^ Hex-encoding of the script or key credential.
-  , _stakingCredentialKind  :: !StakingCredKind
+  , stakingCredentialKind  :: !StakingCredKind
   -- ^ See `StakingCredKind`.
-  , _stakingCredentialPointer  :: !(Maybe ChainPointer)
+  , stakingCredentialPointer  :: !(Maybe ChainPointer)
   -- ^ See `ChainPointer`.
-  , _stakingCredentialRewardAddress :: !(Maybe (Bech32StringOf RewardAddress))
+  , stakingCredentialRewardAddress :: !(Maybe (Bech32StringOf RewardAddress))
   -- ^ See `RewardAddress`.
   }
   deriving stock (Show, Eq, Ord, Generic)
   deriving (FromJSON, ToJSON)
-  via CustomJSON '[FieldLabelModifier '[StripPrefix "_stakingCredential", CamelToSnake]] StakingCredential
+  via CustomJSON '[FieldLabelModifier '[StripPrefix "stakingCredential", CamelToSnake]] StakingCredential
 
 -- | Information decoded from a Cardano address.
 data AddressInfo = AddressInfo
-  { _addressInfoHex         :: !(HexStringOf Address)
+  { addressInfoHex         :: !(HexStringOf Address)
   -- ^ Hexadecimal format encoding of the given address.
-  , _addressInfoBech32      :: !(Maybe (Bech32StringOf Address))
+  , addressInfoBech32      :: !(Maybe (Bech32StringOf Address))
   -- ^ Bech32 representation of the given address. Present for Shelly & stake addresses whereas byron addresses are encoded in Base58.
-  , _addressInfoNetwork     :: !(Maybe NetworkId)
+  , addressInfoNetwork     :: !(Maybe NetworkId)
   -- ^ See `NetworkId`.
-  , _addressInfoPaymentCred :: !(Maybe PaymentCredential)
+  , addressInfoPaymentCred :: !(Maybe PaymentCredential)
   -- ^ See `PaymentCredential`.
-  , _addressInfoStakingCred :: !(Maybe StakingCredential)
+  , addressInfoStakingCred :: !(Maybe StakingCredential)
   -- ^ See `StakingCredential`.
   }
   deriving stock (Show, Eq, Ord, Generic)
   deriving (FromJSON, ToJSON)
-  via CustomJSON '[FieldLabelModifier '[StripPrefix "_addressInfo", CamelToSnake]] AddressInfo
+  via CustomJSON '[FieldLabelModifier '[StripPrefix "addressInfo", CamelToSnake]] AddressInfo
 
 -- | Output reference of an UTxO. This is different from `OutputReference` type as later JSON representation is a string whereas this has an object format.
 data OutputReferenceObject = OutputReferenceObject
-  { _outputReferenceObjectIndex  :: !TxIndex
-  , _outputReferenceObjectTxHash :: !TxHash
+  { outputReferenceObjectIndex  :: !TxIndex
+  , outputReferenceObjectTxHash :: !TxHash
   }
   deriving stock (Show, Eq, Generic)
   deriving (FromJSON, ToJSON)
-  via CustomJSON '[FieldLabelModifier '[StripPrefix "_outputReferenceObject", CamelToSnake]] OutputReferenceObject
+  via CustomJSON '[FieldLabelModifier '[StripPrefix "outputReferenceObject", CamelToSnake]] OutputReferenceObject
 
 -- | UTxO IDs for all the unspent transaction outputs at an address.
 data PaginatedOutputReferenceObject = PaginatedOutputReferenceObject
-  { _paginatedOutputReferenceObjectData        :: ![OutputReferenceObject]
+  { paginatedOutputReferenceObjectData        :: ![OutputReferenceObject]
   -- ^ See `OutputReferenceObject`.
-  , _paginatedOutputReferenceObjectLastUpdated :: !LastUpdated
+  , paginatedOutputReferenceObjectLastUpdated :: !LastUpdated
   -- ^ See `LastUpdated`.
-  , _paginatedOutputReferenceObjectNextCursor  :: !(Maybe NextCursor)
+  , paginatedOutputReferenceObjectNextCursor  :: !(Maybe NextCursor)
   -- ^ See `NextCursor`.
   }
   deriving stock (Show, Eq, Generic)
   deriving (FromJSON, ToJSON)
-  via CustomJSON '[FieldLabelModifier '[StripPrefix "_paginatedOutputReferenceObject", CamelToSnake]] PaginatedOutputReferenceObject
+  via CustomJSON '[FieldLabelModifier '[StripPrefix "paginatedOutputReferenceObject", CamelToSnake]] PaginatedOutputReferenceObject
 
 instance IsTimestamped PaginatedOutputReferenceObject where
   type TimestampedData PaginatedOutputReferenceObject = [OutputReferenceObject]
-  getTimestampedData = _paginatedOutputReferenceObjectData
-  getTimestamp = _paginatedOutputReferenceObjectLastUpdated
+  getTimestampedData = paginatedOutputReferenceObjectData
+  getTimestamp = paginatedOutputReferenceObjectLastUpdated
 
 instance HasCursor PaginatedOutputReferenceObject where
-  getNextCursor = _paginatedOutputReferenceObjectNextCursor
+  getNextCursor = paginatedOutputReferenceObjectNextCursor