+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-# Revision history for maestro
+# Revision history for `maestro-sdk`
 
-## 0.1.0.0 -- YYYY-mm-dd
+## [1.1.0](https://github.com/maestro-org/haskell-sdk/compare/v1.0.0..v1.1.0) -- 2023-09-23
 
-* First version. Released on an unsuspecting world.
+* Support of v0 family of endpoints have been dropped, [#33](https://github.com/maestro-org/haskell-sdk/pull/33).
+* Support of backoff, to automatically handle rate limit errors by allowing clients to use exponential backoff with maximum delay threshold, [#40](https://github.com/maestro-org/haskell-sdk/pull/40).
+* Support of Preview network is added, [#37](https://github.com/maestro-org/haskell-sdk/pull/37).